Improve request error handling to return identifier when it's possible.
The identifier in a JSON RPC 2.0 request is used by the user to recognize which
 request has been sent and adapt his behaviour.
On parsing error, there was no identifier returned, that was a problem because
it's possible to retrieve the identifier but no method or version.
By returning the identifier even on parsing error (when it's possible), it
helps the user to correctly manage his request.

Resolves: #13
Signed-off-by: Vasek - Tom C <tom.chauveau@epitech.eu>